requirements/base.txt
Django==3.2.19
django-admin-sortable==2.2.4
django-anymail==9.0
django-autoslug==1.9.8
django-avatar==7.0.1
django-bootstrap3==22.2
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-filter==22.1
django-fontawesome-5==1.0.18
django-geoposition-2==0.4.0
django-pandas==0.6.6
django-redis==5.2.0
django-select2==8.1.0
django-simple-history==3.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
google-api-python-client==1.12.11
pandas==1.5.3
Pillow==9.4.0
Pyrebase4==4.4.3
python-slugify==8.0.0
redis==4.5.4
requests==2.31.0
git+https://github.com/Benbb96/soundcloud-python.git#egg=soundcloud
spotipy==2.23.0
